post component and page grid addressing
