+++ src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import AppBar from './components/AppBar';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
-import AllPosts from './pages/AllPosts/index';
+import AllPosts from './pages/AllPosts';
 
 export default function App() {
   return (
