+++ src/components/AppFilterMenu.js
@@ -3,8 +3,17 @@ import Button from '@material-ui/core/Button';
 import Menu from '@material-ui/core/Menu';
 import MenuItem from '@material-ui/core/MenuItem';
 import Grid from '@material-ui/core/Grid';
+import { makeStyles } from '@material-ui/core';
+
+const useStyles = makeStyles(() => ({
+  root: {
+    textAlign: 'center'
+  }
+}));
 
 export default function SimpleMenu() {
+  const classes = useStyles();
+
   const [anchorEl, setAnchorEl] = React.useState(null);
 
   const handleClick = (event) => {
@@ -16,33 +25,33 @@ export default function SimpleMenu() {
   };
 
   return (
-    <Grid
-      container
-      spacing={3}
-      direction="column"
-      justify="center"
-      alignItems="center"
-    >
-      <Grid item xs={12} sm={3}>
-        <Button
-          aria-controls="simple-menu"
-          aria-haspopup="true"
-          onClick={handleClick}
-        >
-          Filter Apps
-        </Button>
-        <Menu
-          id="simple-menu"
-          anchorEl={anchorEl}
-          keepMounted
-          open={Boolean(anchorEl)}
-          onClose={handleClose}
-        >
-          <MenuItem onClick={handleClose}>Field</MenuItem>
-          <MenuItem onClick={handleClose}>Age</MenuItem>
-          <MenuItem onClick={handleClose}>User</MenuItem>
-        </Menu>
-      </Grid>
+    <Grid item xs={12} className={classes.root}>
+      <Button
+        aria-controls="filter-menu"
+        aria-haspopup="true"
+        onClick={handleClick}
+      >
+        Filter Apps
+      </Button>
+      <Menu
+        id="filter-menu"
+        anchorEl={anchorEl}
+        open={Boolean(anchorEl)}
+        onClose={handleClose}
+        anchorOrigin={{
+          vertical: 'bottom',
+          horizontal: 'center'
+        }}
+        keepMounted
+        transformOrigin={{
+          vertical: 'top',
+          horizontal: 'center'
+        }}
+      >
+        <MenuItem onClick={handleClose}>Field</MenuItem>
+        <MenuItem onClick={handleClose}>Age</MenuItem>
+        <MenuItem onClick={handleClose}>User</MenuItem>
+      </Menu>
     </Grid>
   );
 }
